image: T5195: vyos.util -> vyos.utils package refactoring
